+++ tests/test_tools_ids_update.py
@@ -1,7 +1,7 @@
 from langworld_db_data.tools.common.files.csv_xls import (
     remove_rows_with_given_content_in_lookup_column,
 )
-from langworld_db_data.tools.ids.update import update_indices_after_given_line_number_if_necessary
+from langworld_db_data.tools.common.ids import update_indices_after_given_line_number_if_necessary
 
 
 def test_update_indices_after_given_line_number_if_necessary_in_features_update_is_necessary(
